# Conflicts:
#	src/main/java/me/vse/fintrackserver/configs/CorsConfig.java
#	src/main/java/me/vse/fintrackserver/controller/AccountController.java
#	src/main/java/me/vse/fintrackserver/controller/UserController.java
#	src/main/java/me/vse/fintrackserver/model/Group.java
#	src/main/java/me/vse/fintrackserver/services/AccountService.java
#	src/main/java/me/vse/fintrackserver/services/GroupService.java
#	src/main/resources/application-prod.properties
#	src/main/resources/liquibase/dbchangelog.xml

+++ src/main/java/me/vse/fintrackserver/rest/requests/TransactionRequest.java
@@ -12,15 +12,13 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 public class TransactionRequest {
-
-
     private String id;
     private String accountId;
     private String forAssetId;
     private String receiverId;
     private String categoryId;
     private TransactionTypes type;
-    private Long amount;
+    private double amount;
     private LocalDateTime executionDateTime;
     private String note;
     private Double lat;
